Create actividad and chages of reporte

+++ src/main/java/ryuulogic/adatics/reportes/ReporteRepository.java
@@ -2,5 +2,5 @@ package ryuulogic.adatics.reportes;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ReporteRepository extends CrudRepository<Reporte, Integer> {
+public interface ReporteRepository extends CrudRepository<Reporte, Long> {
 }
